At this point I forget but it seems like progress was made

+++ install/arch.sh
@@ -7,14 +7,14 @@ if ! ping -c 1 google.com > /dev/null 2>&1; then
 fi
 
 install () {
-    sudo pacman -Sq --noconfirm $@
+    sudo pacman -S $@
 }
 
 
 # git
 echo "configure git? [Y/n]"
-read -n 1 config_git
-if [ -z "$config_git" ] || [ "$config_git" = "Y" ] || [ "$config_git" = 'y' ]; then
+read -n 1 config_section
+if [ -z "$config_section" ] || [ "$config_section" = "Y" ] || [ "$config_section" = 'y' ]; then
     ssh-keygen -t rsa -b 4096 -C "kyle.thompson228@gmail.com"
     ssh-add ~/.ssh/id_rsa
     echo "add id_rsa.pub to github and press [ENTER]"
@@ -33,43 +33,64 @@ git clone --quiet git@github.com:Kyle-Thompson/dotfiles.git ~/.dotfiles
 
 
 # media  TODO: Host media somewhere and fetch it here
-curl -s http://i.imgur.com/SpBfUZi.jpg --create-dirs -o ~/Pictures/Wallpapers/deer.jpg
+echo "Fetch media? [Y/n]"
+read -n 1 config_section
+if [ -z "$config_section" ] || [ "$config_section" = "Y" ] || [ "$config_section" = 'y' ]; then
+	curl -s http://i.imgur.com/SpBfUZi.jpg --create-dirs -o ~/media/wallpapers/deer.jpg
+fi
 
 
 # neovim
-install neovim
-install python2-pip python-pip
-pip2 install --upgrade pip neovim
-pip install --upgrade pip neovim
-rm -rf ~/.config/nvim
-ln -s ~/.dotfiles/nvim ~/.config/nvim
+echo "configure neovim? [Y/n]"
+read -n 1 config_section
+if [ -z "$config_section" ] || [ "$config_section" = "Y" ] || [ "$config_section" = 'y' ]; then
+	install neovim
+	install python2-pip python-pip
+	pip2 install --upgrade pip neovim
+	pip install --upgrade pip neovim
+	rm -rf ~/.config/nvim
+	ln -s ~/.dotfiles/nvim ~/.config/nvim
+fi
 
 
 # general WM
-install feh thunar
+echo "configure WMs? [Y/n]"
+read -n 1 config_section
+if [ -z "$config_section" ] || [ "$config_section" = "Y" ] || [ "$config_section" = 'y' ]; then
+	install feh 
+	install thunar
 
-# i3
-install i3
-rm -rf ~/.config/i3
-#chmod +x ~/.dotfiles/etc/monitor-layout.sh
-ln -s ~/.dotfiles/i3 ~/.config/i3
+	# i3
+	install i3
+	rm -rf ~/.config/i3
+	#chmod +x ~/.dotfiles/etc/monitor-layout.sh
+	ln -s ~/.dotfiles/i3 ~/.config/i3
 
-# openbox
-install openbox
-rm -rf ~/.config/openbox
-ln -s ~/.dotfiles/openbox ~/.config/openbox
+	# openbox
+	install openbox
+	rm -rf ~/.config/openbox
+	ln -s ~/.dotfiles/openbox ~/.config/openbox
+fi
 
 
 # zsh
-install zsh
-chsh -s `which zsh`
-rm ~/.zshrc
-ln -s ~/.dotfiles/zsh/zshrc ~/.zshrc
+echo "configure zsh? [Y/n]"
+read -n 1 config_section
+if [ -z "$config_section" ] || [ "$config_section" = "Y" ] || [ "$config_section" = 'y' ]; then
+	install zsh
+	chsh -s `which zsh`
+	rm ~/.zshrc
+	ln -s ~/.dotfiles/zsh/zshrc ~/.zshrc
+fi
 
 
 # termite
-install termite
-rm -rf ~/.config/termite
-ln -s ~/.dotfiles/termite ~/.config/termite
+echo "configure termite? [Y/n]"
+read -n 1 config_section
+if [ -z "$config_section" ] || [ "$config_section" = "Y" ] || [ "$config_section" = 'y' ]; then
+	install termite
+	rm -rf ~/.config/termite
+	ln -s ~/.dotfiles/termite ~/.config/termite
+fi
